+++ easybatch-tutorials/src/main/java/io/github/benas/easybatch/tutorials/recipes/RecipeRecord.java
@@ -17,7 +17,7 @@ public class RecipeRecord implements Record<Recipe> {
     }
 
     @Override
-    public long getNumber() {
+    public int getNumber() {
         return number;
     }
